All tests fixed
