+++ lib/msal-browser/test/encode/Base64Decode.spec.ts
@@ -67,7 +67,7 @@ describe("Base64Decode.ts Unit Tests", () => {
                 "ver": "2.0",
                 "iss": `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
                 "sub": "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                "exp": "1536361411",
+                "exp": 1536361411,
                 "name": "Abe Lincoln",
                 "preferred_username": "AbeLi@microsoft.com",
                 "oid": "00000000-0000-0000-66f3-3332eca7ea81",
