+++ lib/msal-browser/test/error/BrowserConfigurationAuthError.spec.ts
@@ -31,30 +31,4 @@ describe("BrowserConfigurationAuthError Unit Tests", () => {
         expect(err.name).to.equal("BrowserConfigurationAuthError");
         expect(err.stack).to.include("BrowserConfigurationAuthError.spec.ts");
     });
-
-    it("createInvalidCallbackObjectError()", () => {
-        const err: BrowserConfigurationAuthError = BrowserConfigurationAuthError.createInvalidCallbackObjectError(null);
-
-        expect(err instanceof BrowserConfigurationAuthError).to.be.true;
-        expect(err instanceof AuthError).to.be.true;
-        expect(err instanceof Error).to.be.true;
-        expect(err.errorCode).to.equal(BrowserConfigurationAuthErrorMessage.invalidCallbackObject.code);
-        expect(err.errorMessage).to.include(BrowserConfigurationAuthErrorMessage.invalidCallbackObject.desc);
-        expect(err.message).to.include(BrowserConfigurationAuthErrorMessage.invalidCallbackObject.desc);
-        expect(err.name).to.equal("BrowserConfigurationAuthError");
-        expect(err.stack).to.include("BrowserConfigurationAuthError.spec.ts");
-    });
-
-    it("createRedirectCallbacksNotSetError()", () => {
-        const err: BrowserConfigurationAuthError = BrowserConfigurationAuthError.createRedirectCallbacksNotSetError();
-
-        expect(err instanceof BrowserConfigurationAuthError).to.be.true;
-        expect(err instanceof AuthError).to.be.true;
-        expect(err instanceof Error).to.be.true;
-        expect(err.errorCode).to.equal(BrowserConfigurationAuthErrorMessage.noRedirectCallbacksSet.code);
-        expect(err.errorMessage).to.include(BrowserConfigurationAuthErrorMessage.noRedirectCallbacksSet.desc);
-        expect(err.message).to.include(BrowserConfigurationAuthErrorMessage.noRedirectCallbacksSet.desc);
-        expect(err.name).to.equal("BrowserConfigurationAuthError");
-        expect(err.stack).to.include("BrowserConfigurationAuthError.spec.ts");
-    });
 });
